deploy/requirements.txt
click==8.1.7
colorama==0.4.6
gitdb==4.0.11
GitPython==3.1.43
python-dotenv==1.0.1
setuptools==72.2.0
smmap==5.0.1
